Move diagonalization function to polyhedron.
