+++ coxeter/shape_classes/base_classes.py
@@ -8,8 +8,6 @@ parts of the code base that either require or return shapes.
 
 from abc import ABC, abstractmethod
 
-import numpy as np
-
 
 class Shape(ABC):
     """An abstract representation of a shape in N dimensions."""
@@ -75,13 +73,3 @@ class Shape3D(Shape):
     def inertia_tensor(self):
         """:math:`(3, 3)` :class:`numpy.ndarray`: Get the inertia tensor."""
         pass
-
-    def diagonalize_inertia(self):
-        """Orient the shape along its principal axes.
-
-        The principal axes of a shape are defined by the eigenvectors of the inertia
-        tensor. This method computes the inertia tensor of the shape, diagonalizes it,
-        and then rotates the shape by the corresponding orthogonal transformation.
-        """
-        principal_moments, principal_axes = np.linalg.eigh(self.inertia_tensor)
-        self._vertices = np.dot(self._vertices, principal_axes)
